# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API KosKuy PAA",
        "description": "Dokumentasi Swagger UI untuk testing endpoint API Laravel 12",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/facilities": {
            "get": {
                "tags": [
                    "Facilities"
                ],
                "summary": "List fasilitas",
                "description": "Access: Semua user login",
                "operationId": "listFacilities",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Facilities"
                ],
                "summary": "Buat fasilitas baru",
                "description": "Access: Pemilik Kos only",
                "operationId": "createFacility",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Parkir"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/facilities/{id}": {
            "get": {
                "tags": [
                    "Facilities"
                ],
                "summary": "Detail fasilitas",
                "description": "Access: Semua user login. Menampilkan detail satu fasilitas.",
                "operationId": "showFacility",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Facilities"
                ],
                "summary": "Update fasilitas",
                "description": "Access: Pemilik Kos only",
                "operationId": "updateFacility",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "WiFi"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Facilities"
                ],
                "summary": "Hapus fasilitas",
                "description": "Access: Pemilik Kos only",
                "operationId": "deleteFacility",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Ambil profil",
                "description": "Access: Semua user login. Menampilkan profil user yang sedang login.",
                "operationId": "showProfile",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update profil",
                "description": "Access: Semua user login. Mengubah nama, email, nomor HP, dan alamat user yang sedang login.",
                "operationId": "updateProfile",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Kiarra"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "kiarra@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "08123456789",
                                        "nullable": true
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Jl. Mawar No. 10",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/profile/password": {
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Ganti password",
                "description": "Access: Semua user login. Mengganti password user yang sedang login.",
                "operationId": "updateProfilePassword",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "current_password",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "current_password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "passwordbaru123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "passwordbaru123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/profile/photo": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Upload foto profil",
                "description": "Access: Semua user login. Mengunggah atau mengganti foto profil user yang sedang login.",
                "operationId": "uploadProfilePhoto",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "profile_picture"
                                ],
                                "properties": {
                                    "profile_picture": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Uploaded"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter token in format (Bearer <token>)"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}